requirements.txt
numpy
pandas
matplotlib
pillow
rasterio
geopandas
shapely
torchmetrics
ipykernel
jupyter